requirements.txt
keras
numpy
djangorestframework
matplotlib
jieba
pandas
sqlalchemy
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==2.2.7
idna==2.9
mysqlclient==1.4.6
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
